refactor alsa thread to use alsa-device
