+++ src/sound/jam_socket.rs
@@ -95,7 +95,6 @@ mod test_jam_socket {
     #[test]
     fn build_socket() {
         let sock = JamSocket::new(19990).unwrap();
-        println!("sock: {}", sock);
         assert!(true);
     }
     #[test]
